requirements.txt
openai
gradio
mysql-connector-python
python-dotenv
